chart-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyswisseph==2.10.3.2
pytz==2024.1
